Add API endpoints
Added professor and course endpoint (grade_distribution commented out)
Moved loop through section ids to index.tsx
